requirements.txt
pyautogui
python-dotenv
opencv-python
pillow